+++ src/main/java/de/is24/deadcode4j/analyzer/JsfAnnotationsAnalyzer.java
@@ -11,7 +11,16 @@ import de.is24.deadcode4j.Analyzer;
 public final class JsfAnnotationsAnalyzer extends AnnotationsAnalyzer implements Analyzer {
 
     public JsfAnnotationsAnalyzer() {
-        super("_JSF-Annotation_", "javax.faces.convert.FacesConverter");
+        super("_JSF-Annotation_",
+                "javax.faces.component.behavior.FacesBehavior",
+                "javax.faces.convert.FacesConverter",
+                "javax.faces.event.ListenerFor",
+                "javax.faces.event.ListenersFor",
+                "javax.faces.event.NamedEvent",
+                "javax.faces.render.FacesBehaviorRenderer",
+                "javax.faces.render.FacesRenderer",
+                "javax.faces.validator.FacesValidator",
+                "javax.faces.view.facelets.FaceletsResourceResolver");
     }
 
 }